+++ requirements.txt
@@ -1,5 +1,15 @@
-streamlit
-shapely
-numpy
-pyproj
-httpx
+streamlit==1.40.1
+shapely==2.0.6
+numpy==2.1.3
+pyproj==3.7.0
+httpx==0.27.2
+folium==0.18.0
+streamlit-folium==0.23.2
+pydeck==0.9.1
+pandas==2.2.3
+matplotlib==3.9.2
+pykml==0.2.0
+geopy==2.4.1
+scikit-learn==1.5.2
+contextily==1.6.2
+geopandas==1.0.1
